DAOS-14251 control: Allow premounted empty tmpfs
In the special case where the tmpfs has already been
mounted but is empty, don't skip NVMe format and
configuration. Enables the use case of running
daos_server in a container with an external tmpfs.

Deprecates --recreate-superblocks flag and removes
related logic that is no longer needed due to the
new support for existing tmpfs.

Required-githooks: true

Change-Id: Iaa4255c45cf7f3eaf3cb244351c99f09072a43e6
Signed-off-by: Michael MacDonald <mjmac@google.com>